Accept integers in programmatically created filter/predicate arguments. Issue no : 296
Modified eskip to parse numbers as int64 too.
Some filters and predicates relying this accept only float64 args in number positions.
This is very inconvenient when creating filters programmatically.
Where possible fixed these filters to accepts integers (int64), too.

+++ eskip/eskip_test.go
@@ -302,7 +302,7 @@ func TestParseFilters(t *testing.T) {
 	}, {
 		"success",
 		`filter1(3.14) -> filter2("key", 42)`,
-		[]*Filter{{Name: "filter1", Args: []interface{}{3.14}}, {Name: "filter2", Args: []interface{}{"key", float64(42)}}},
+		[]*Filter{{Name: "filter1", Args: []interface{}{3.14}}, {Name: "filter2", Args: []interface{}{"key", int64(42)}}},
 		false,
 	}} {
 		fs, err := ParseFilters(ti.expression)
